# adodasopenapi.yaml
openapi: 3.0.3
info:
  title: Adodas API
  description: |
    API REST completa para gestión de **clientes, ítems, pedidos, reembolsos y servicios**.
    
    ## ✨ Características:
    - ✅ **CRUD completo** todos los recursos
    - ✅ **Validaciones robustas** (DNI, emails, precios)
    - ✅ **Filtros avanzados** por estado, precio, cliente
    - ✅ **Errores específicos** 400/404/500 diferenciados
    - ✅ **Ejemplos reales** en todas las respuestas
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Desarrollo Local
...
components:
  schemas:
    ErrorResponseDto:
      type: object
      properties:
        code:
          type: integer
          description: Código HTTP del error
        message:
          type: string
          description: Descripción del error
      required: [ code, message ]
...
    OrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 101
        customerId:
          type: integer
          format: int64
          example: 12
        items:
          type: array
          description: Lista de IDs de ítems incluidos en el pedido
          items:
            type: integer
            format: int64
          example: [ 3, 14, 22 ]
        orderPrice:
          type: number
          format: float
          minimum: 0
          example: 79.85
        delivered:
          type: boolean
          example: false
    RefundDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 25
        reason:
          type: string
          description: Motivo del reembolso.
          example: "Producto defectuoso"
        refundDate:
          type: string
          format: date
          description: Fecha en la que se realizó el reembolso.
          example: "2026-01-28"
        amount:
          type: number
          format: float
          minimum: 0
          description: Importe total del reembolso.
          example: 49.99
        orderId:
          type: integer
          format: int64
          description: ID del pedido relacionado.
          example: 320
        customerId:
          type: integer
          format: int64
          description: ID del cliente que recibe el reembolso.
          example: 8
        approved:
          type: boolean
          description: Indica si el reembolso fue aprobado.
          example: true
      required: [ reason, refundDate, amount ]
...
paths:
...
  /orders:
    get:
      tags: [ ORDERS ]
      summary: Lista todos los pedidos
      description: Obtiene pedidos con filtros opcionales por estado de entrega o rango de precio.
      parameters:
        - name: delivered
          in: query
          schema:
            type: boolean
          description: Filtra por pedidos entregados o pendientes.
        - name: minPrice
          in: query
          schema:
            type: number
            format: float
            minimum: 0
          description: Precio mínimo de los pedidos.
        - name: maxPrice
          in: query
          schema:
            type: number
            format: float
            minimum: 0
          description: Precio máximo de los pedidos.
      responses:
        '200':
          description: Lista de pedidos obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ ORDERS ]
      summary: Crea un nuevo pedido
      description: Registra un nuevo pedido con validaciones de total y cliente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
      responses:
        '201':
          description: Pedido creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /orders/{orderId}:
    put:
      tags: [ ORDERS ]
      summary: Modifica un pedido existente
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del pedido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
      responses:
        '200':
          description: Pedido modificado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Pedido no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Pedido no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    patch:
      tags: [ ORDERS ]
      summary: Actualiza parcialmente un pedido
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del pedido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                delivered: true
                totalPrice: 85.55
      responses:
        '200':
          description: Pedido actualizado parcialmente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Error de validación o cuerpo inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Pedido no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Pedido no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ ORDERS ]
      summary: Elimina un pedido
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del pedido a eliminar.
      responses:
        '204':
          description: Pedido eliminado correctamente (sin contenido en la respuesta).
        '404':
          description: Pedido no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Pedido no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /refunds:
    get:
      tags: [ REFUNDS ]
      summary: Lista todos los reembolsos
      description: Obtiene una lista de reembolsos aplicando filtros opcionales por motivo, monto mínimo y estado de aprobación.
      parameters:
        - name: reason
          in: query
          required: false
          schema:
            type: string
          description: Filtra por motivo del reembolso.
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
          description: Filtra por importe mínimo.
        - name: isApproved
          in: query
          required: false
          schema:
            type: boolean
          description: Filtra por si el reembolso fue aprobado.
      responses:
        '200':
          description: Lista de reembolsos obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefundDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ REFUNDS ]
      summary: Crea un nuevo reembolso
      description: Registra un nuevo reembolso validando motivo, cantidad y fecha.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundDto'
      responses:
        '201':
          description: Reembolso creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /refunds/{refundId}:
    put:
      tags: [ REFUNDS ]
      summary: Modifica un reembolso existente
      parameters:
        - name: refundId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del reembolso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundDto'
      responses:
        '200':
          description: Reembolso modificado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundDto'
        '400':
          description: Error de validación en los datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Reembolso no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Reembolso no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    patch:
      tags: [ REFUNDS ]
      summary: Actualiza parcialmente un reembolso
      parameters:
        - name: refundId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del reembolso a actualizar parcialmente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                amount: 42.80
                approved: true
      responses:
        '200':
          description: Reembolso actualizado parcialmente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundDto'
        '400':
          description: Error de validación o cuerpo inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Reembolso no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Reembolso no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ REFUNDS ]
      summary: Elimina un reembolso
      parameters:
        - name: refundId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del reembolso a eliminar.
      responses:
        '204':
          description: Reembolso eliminado correctamente (sin contenido en la respuesta).
        '404':
          description: Reembolso no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Reembolso no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
...